test: Add tests for date identification
This change moves the date identification functionality (DateFinder, etc) into FileawayCore and adds dedicated tests for date identification. It also enables FileawayCore tests on GitHub Actions.
