+++ core/FileawayCore/TitleFinder.swift
@@ -20,29 +20,6 @@
 
 import Foundation
 
-class DateFinder {
-
-    static var cache: NSCache<NSString, NSArray> = {
-        return NSCache()
-    }()
-
-    static var dateFormatter: DateFormatter = {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyyy-MM-dd"
-        return dateFormatter
-    }()
-
-    static func dates(from string: String) -> Array<Date> {
-        if let dates = Self.cache.object(forKey: string as NSString) as? Array<Date> {
-            return dates
-        }
-        let dates = string.tokens.compactMap { Self.dateFormatter.date(from: $0) }
-        Self.cache.setObject(dates as NSArray, forKey: string as NSString)
-        return dates
-    }
-
-}
-
 class TitleFinder {
 
     static var cache: NSCache<NSString, NSString> = {
@@ -64,35 +41,4 @@ class TitleFinder {
         Self.cache.setObject(title as NSString, forKey: string as NSString)
         return title
     }
-}
-
-class FileInfo: Identifiable, Hashable {
-
-    public var id: URL { url }
-    let url: URL
-    let name: String
-    var date: Date?
-    var sortDate: Date {
-        date ?? Date.distantPast
-    }
-
-    let directoryUrl: URL
-
-    init(url: URL) {
-        self.url = url
-        let name = url.lastPathComponent.deletingPathExtension
-        self.date = DateFinder.dates(from: name).first
-        let title = TitleFinder.title(from: name)
-        self.name = title.isEmpty ? name : title
-        self.directoryUrl = url.deletingLastPathComponent()
-    }
-
-    static func == (lhs: FileInfo, rhs: FileInfo) -> Bool {
-        return lhs.url == rhs.url
-    }
-
-    func hash(into hasher: inout Hasher) {
-        hasher.combine(url)
-    }
-
 }
