+++ macos/Fileaway/Views/FileRow.swift
@@ -20,6 +20,7 @@
 
 import SwiftUI
 
+import FileawayCore
 import Interact
 
 struct FileRow: View {
